% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDJM_wrapper.R
\docType{data}
\name{control_list}
\alias{control_list}
\title{control_list}
\description{
This list contains a list of parameters specifying the joint model.
}
\details{
\itemize{
  \item ID_name the variable name for the patient ID in both
  longitudinal data and survival data.
  \item item_name the variable name for the longitudinal outcomes
  in the longitudinal data.
  \item value_name the variable name for the longitudinal measurements
  in the longitudinal data.
  \item time_name the variable name for the measurement timepoints in the
  longitudinal data.
  \item fix_cov a set of variables names indicating the covariates of
  fixed-effects in the longitudinal submodel.
  If NULL, not baseline covariates are included.
  \item random_cov a set of variables names indicating the covariates of
  random-effects in the longitudinal submodel.
  If NULL, not baseline covariates are included.
  \item FUN a function specifying the time-related basis functions in
  the longitudinal submodel.
  \item ran_time_ind a vector of integers specifying which
  time-related basis functions are also included with random-effects in
  the longitudinal submodel.
  \item surv_time_name the variable name for the survival time
  in the survival data.
  \item surv_status_name the variable name for the censoring indicator
  in the survival data.
  \item surv_cov a set of variables names specifying the baseline covariates
  in the survival submodel.
  \item n_points an integer indicating the numebr of nodes being used in
  the Gaussian quadrature.
}
}
\author{
Jiehuan Sun \email{jiehuan.sun@gmail.com}
}
\keyword{data}
