% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDJM_wrapper.R
\docType{data}
\name{LongData}
\alias{LongData}
\title{Simulated Longtidunal Data}
\format{
A data frame with 48700 rows and 4 variables
}
\usage{
data(HDJMdata)
}
\description{
This dataset contains longitudinal outcomes.
}
\details{
\itemize{
  \item ID patient ID
  \item item types of longitudinal outcome
  \item years measurement timepoints
  \item value measurements
}
}
\author{
Jiehuan Sun \email{jiehuan.sun@gmail.com}
}
\keyword{data}
