% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HDJM_wrapper.R
\docType{data}
\name{SurvData}
\alias{SurvData}
\title{Simulated Survival Data}
\format{
A data frame with 100 rows and 4 variables
}
\usage{
data(HDJMdata)
}
\description{
This dataset contains survival outcome.
}
\details{
\itemize{
  \item ID patient ID
  \item fstat censoring indicator
  \item ftime survival time
  \item x baseline covariates
}
}
\author{
Jiehuan Sun \email{jiehuan.sun@gmail.com}
}
\keyword{data}
